when rect edge is dragged off the right axis boundary, use the left axis boundary to calculate the max width
The rects left boundary was being used here, but this only works for the major rects before the minor rects are added. Once the minor rects are added, the major rects left boundary is updated to be the minor rects right boundary. When a rects right edge is dragged out of the axis towards the right, this represents creating a max width rect, so we should use the axis left boundary to compute this max width